Private API

Here, all non-exported types and methods are listed, which the average user does not need when using the package.

Peridynamics.AbstractBC

Peridynamics.AbstractIC

Peridynamics.AbstractPDAnalysis

Peridynamics.AbstractPDMaterial

Peridynamics.AbstractTimeDiscretization

Peridynamics.BondBasedBody

Peridynamics.init_body

Peridynamics.sphere_radius

Types

Peridynamics.AbstractIC — Type

AbstractIC

Abstract type for initial conditions.

Peridynamics.AbstractBC — Type

AbstractBC

Abstract type for boundary conditions.

Peridynamics.AbstractPDAnalysis — Type

AbstractPDAnalysis

Abstract type for a peridynamic analysis.

Peridynamics.AbstractPDMaterial — Type

AbstractPDMaterial

Abstract type for a peridynamic material.

Peridynamics.AbstractTimeDiscretization — Type

AbstractTimeDiscretization

Abstract type for time discretizations.

Peridynamics.BondBasedBody — Type

BondBasedBody

Simulation body that contains all needed informations to run a simulation with bond-based peridynamics.

Fields

n_points::Int:

n_bonds::Int:

n_threads::Int:

unique_bonds::Bool:

owned_points::Vector{UnitRange{Int}}:

owned_bonds::Vector{UnitRange{Int}}:

single_tids::Vector{Tuple{Int, Int}}:

multi_tids::Vector{Tuple{Int, Vector{Int}}}:

bond_data::Vector{Tuple{Int, Int, Float64, Bool}}:

volume::Vector{Float64}:

cells::Vector{MeshCell{VTKCellType, Tuple{Int64}}}:

n_family_members::Vector{Int}:

n_active_family_members::Matrix{Int}:

position::Matrix{Float64}:

displacement::Matrix{Float64}:

velocity::Matrix{Float64}:

velocity_half::Matrix{Float64}:

acceleration::Matrix{Float64}:

b_int::Array{Float64, 3}:

b_ext::Matrix{Float64}:

damage::Vector{Float64}:

bond_failure::Vector{Int}:

BondBasedBody(mat::PDMaterial{BondBasedMaterial}, pc::PointCloud)

Arguments:

mat::PDMaterial{BondBasedMaterial}:

pc::PointCloud:

Functions

Peridynamics.init_body — Function

init_body(mat::PDMaterial, pc::PointCloud)

Material interface function for initialization of the body <: AbstractPDBody used for the simulation.

Peridynamics.sphere_radius — Function

sphere_radius(vol::T) where {T<:Real}

Calculate the radius $r$ of the sphere by equation

\[r = \sqrt[3]{\frac{3 \; V}{4 \; \pi}}\]

with specified sphere volume $V$.

Arguments

vol::T where {T<:Real}: volume $V$ of the sphere

Returns

Float64: radius $r$ of sphere with volume vol