Public API

Here, all exported types and methods are listed, which the user needs when using the package.

Peridynamics.BodySetup

Peridynamics.BondBasedMaterial

Peridynamics.Contact

Peridynamics.ContinuumBasedMaterial

Peridynamics.DynamicRelaxation

Peridynamics.ExportSettings

Peridynamics.ForceDensityBC

Peridynamics.MultiMaterial

Peridynamics.PDContactAnalysis

Peridynamics.PDSingleBodyAnalysis

Peridynamics.PointCloud

Peridynamics.PosDepVelBC

Peridynamics.PreCrack

Peridynamics.VelocityBC

Peridynamics.VelocityIC

Peridynamics.VelocityVerlet

Peridynamics.AbaqusMeshConverter.read_inp

Peridynamics.VtkReader.read_vtk

Peridynamics.pcmerge

Peridynamics.submit

Types

Peridynamics.PointCloud — Type

PointCloud

Peridynamic spatial discretization with material points defining a point cloud.

Fields

n_points::Int: number of material points

position::Matrix{Float64}: coordinates of points in reference configuration

volume::Vector{Float64}: material point volumes

failure_flag::BitVector: if failure of point is possible: element=true

radius::Vector{Float64}: radius of the material point sphere

PointCloud(position, volume[, point_sets])

Create a PointCloud by specifying position and volume of all material points. The radius of the point sphere is derived from the volume with the function sphere_radius and the failure_flag set to true for all points.

Arguments

position::Matrix{Float64}: the position of the material points ($3 \times N$-matrix for $N$ material points)

volume::Vector{Float64}: the volume of the material points

point_sets::Dict{String,Vector{Int}}=Dict{String,Vector{Int}}(): optional point sets

Returns

PointCloud: point cloud with specified material point position and volume

Examples

Creating a PointCloud with 4 manually defined points:

julia> position = [
           0.0 1.0 0.0 0.0
           0.0 0.0 1.0 0.0
           0.0 0.0 0.0 1.0
       ]
3×4 Matrix{Float64}:
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0

julia> volume = [1.0, 1.0, 1.0, 1.0]
4-element Vector{Float64}:
 1.0
 1.0
 1.0
 1.0

julia> pc = PointCloud(position, volume)
4-points PointCloud

julia> pc.failure_flag
4-element BitVector:
 1
 1
 1
 1

julia> pc.radius
4-element Vector{Float64}:
 0.6203504908994001
 0.6203504908994001
 0.6203504908994001
 0.6203504908994001

PointCloud(lx, ly, lz, Δx; center_x=0, center_y=0, center_z=0)

Generate a uniformly distributed PointCloud with the lengths lx, ly, lz, and point spacing Δx. Optional keyword arguments provide the possibility to set the center.

Arguments

lx::Real: length of the PointCloud-block in x-direction

ly::Real: length of the PointCloud-block in y-direction

lz::Real: length of the PointCloud-block in z-direction

Δx::Real: point spacing in x-, y- and z- direction

center_x::Real=0: x-coordinate of the PointCloud-block center

center_y::Real=0: y-coordinate of the PointCloud-block center

center_z::Real=0: z-coordinate of the PointCloud-block center

Returns

PointCloud: point cloud with with W, length L, height H and point spacing Δx

Examples

Cube with side length 1 and point spacing $Δx = 0.1$:

julia> PointCloud(1, 1, 1, 0.1)
1000-points PointCloud

Peridynamics.BondBasedMaterial — Type

BondBasedMaterial <: AbstractPDMaterial

Bond based peridynamic material model.

Fields

δ::Float64: horizon

rho::Float64: density

E::Float64: young's modulus

nu::Float64: poisson ratio

G::Float64: shear modulus

K::Float64: bulk modulus

bc::Float64: bond constant

Gc::Float64: critical energy release rate

εc::Float64: critical bond strain

BondBasedMaterial(; horizon::Real, rho::Real, E::Real, [Gc::Real, epsilon_c::Real])

Specify a material with horizon, density rho, Young's modulus E and either the critical energy release rate Gc or the critical bond strain epsilon_c.

Keywords

horizon::Real: horizon

rho::Real:density

E::Real: young's modulus

Gc::Real: critical energy release rate

epsilon_c::Real: critical bond strain

Peridynamics.ContinuumBasedMaterial — Type

BondBasedMaterial <: AbstractPDMaterial

Bond based peridynamic material model.

Fields

δ::Float64: horizon

rho::Float64: density

E::Float64: young's modulus

nu::Float64: poisson ratio

G::Float64: shear modulus

K::Float64: bulk modulus

C1::Float64: parameter for one-neighbor interactions

C2::Float64: parameter for two-neighbor interactions

C3::Float64: parameter for three-neighbor interactions

Gc::Float64: critical energy release rate

εc::Float64: critical bond strain

ContinuumBasedMaterial(kwargs...)

Specify a material with horizon, density rho, Young's modulus E, Poisson ratio nu and either the critical energy release rate Gc or the critical bond strain epsilon_c. The continuum-kinematics-based formulation has in it's core formulas three material parameters $C_1$, $C_2$, and $C_3$. They can be defined by the user, which then overrides the usage of the other material parameters. Be careful if you use this!

Keywords

horizon::Real: horizon

rho::Real:density

E::Real: young's modulus

Gc::Real, optional: critical energy release rate

epsilon_c::Real, optional: critical bond strain

C1::Real, optional: parameter for one-neighbor interactions

C2::Real, optional: parameter for two-neighbor interactions

C3::Real, optional: parameter for three-neighbor interactions

Peridynamics.MultiMaterial — Type

MultiMaterial{T <: AbstractPDMaterial, N, I <: Integer}

Multiple material properties for one body.

Fields

materials::NTuple{N, T}: different materials, all of the same type!

matofpoint::Vector{I}: material index in materials of each point

Example

julia> mat1 = BondBasedMaterial(horizon=1, rho=8e-6, E=2.1e5, Gc=2)
BondBasedMaterial:
  δ:   1.0
  rho: 8.0e-6
  E:   210000.0
  nu:  0.25
  G:   84000.0
  K:   140000.0
  bc:  802140.9131831526
  Gc:  2.0
  εc:  0.0028171808490950554

julia> mat2 = BondBasedMaterial(horizon=1.1, rho=3e-6, E=2.7e4, Gc=1)
BondBasedMaterial:
  δ:   1.1
  rho: 3.0e-6
  E:   27000.0
  nu:  0.25
  G:   10800.0
  K:   18000.0
  bc:  70440.81901751804
  Gc:  1.0
  εc:  0.0052970143846977355

julia> matofpoint = [1, 1, 2, 2] # points 1:2 -> mat1; points 3:4 -> mat2
4-element Vector{Int64}:
 1
 1
 2
 2

julia> mm = MultiMaterial((mat1, mat2), matofpoint)
MultiMaterial{BondBasedMaterial, 2, UInt8} for 4 points

Peridynamics.PreCrack — Type

PreCrack(point_id_set_a::Vector{Int}, point_id_set_b::Vector{Int})

Definition of an preexisting crack in the model. Points in point_id_set_a cannot have interactions with points in point_id_set_b.

Fields

point_id_set_a::Vector{Int}: first point-id set

point_id_set_b::Vector{Int}: second point-id set

Peridynamics.VelocityBC — Type

VelocityBC <: AbstractBC

Velocity boundary condition. The value of the velocity is calculated every time step with the function fun and applied to the dimension dim of the points specified by point_id_set.

Fields

fun::Function: function f(t) with current time t as argument, that calculates the

value of the velocity for each timestep

point_id_set::Vector{Int}: point-id set with all points for which the boundary condition

gets applied every timestep

dim::Int: dimension on which the boundary condition gets applied to. Possible values:

x-direction: dim=1

y-direction: dim=2

z-direction: dim=3

Examples

The constant velocity $v = 0.1$ in $y$-direction gets applied to the first $10$ points:

julia> VelocityBC(t -> 0.1, 1:10, 2)
VelocityBC(var"#7#8"(), [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], 2)

Peridynamics.VelocityIC — Type

VelocityIC <: AbstractIC

Velocity initial condition. The value val of the velocity is applied as initial condition to the dimension dim of the points specified by point_id_set.

Fields

val::Float64: value of the velocity

point_id_set::Vector{Int}: point-id set with all points for which the initial condition

gets applied to

dim::Int: dimension on which the initial condition gets applied to. Possible values:

x-direction: dim=1

y-direction: dim=2

z-direction: dim=3

Peridynamics.PosDepVelBC — Type

PosDepVelBC <: AbstractBC

Position dependent velocity boundary condition. The value of the force density is calculated every time step with the function fun and applied to the dimension dim of the points specified by point_id_set.

Fields

fun::Function: function f(x, y, z, t) with x-, y-, z-position of the point and current

time t as arguments, calculates the value of the force density for each timestep dependent of the point position

point_id_set::Vector{Int}: point-id set with all points for which the boundary condition

gets applied to

dim::Int: dimension on which the initial condition gets applied to. Possible values:

x-direction: dim=1

y-direction: dim=2

z-direction: dim=3

Peridynamics.ForceDensityBC — Type

ForceDensityBC <: AbstractBC

Force density boundary condition. The value of the force density is calculated every time step with the function fun and applied to the dimension dim of the points specified by point_id_set.

Fields

fun::Function: function f(t) with current time t as argument, that calculates the

value of the force density for each timestep

point_id_set::Vector{Int}: point-id set with all points for which the boundary condition

gets applied every timestep

dim::Int: dimension on which the boundary condition gets applied to. Possible values:

x-direction: dim=1

y-direction: dim=2

z-direction: dim=3

Peridynamics.VelocityVerlet — Type

VelocityVerlet <: AbstractTimeDiscretization

Velocity Verlet algorithm for dynamic simulations.

Fields

n_steps::Int: number of time steps

Δt::Float64: critical stable time step

safety_factor::Float64: safety factor for time step

VelocityVerlet(n_steps::Int, Δt::Real = -1; safety_factor::Real = 0.7)

Only the number of time steps n_steps is needed. Specification of Δt and the safety factor are optional.

Peridynamics.DynamicRelaxation — Type

DynamicRelaxation <: AbstractTimeDiscretization

Dynamic Relaxation algorithm for quasi-static simulations.

Fields

n_steps::Int: number of time steps

Δt::Float64: time step

Λ::Float64: damping factor for the mass matrix

DynamicRelaxation(n_steps::Int, Δt::Real = 1; damping_factor::Real = 1)

Only the number of time steps n_steps is needed. Specification of Δt and the damping factor are optional.

Peridynamics.ExportSettings — Type

ExportSettings

Export settings.

Fields

path::String: path where results will be saved

exportfreq::Int: export frequency, will export every exportfreq-th timestep

resultfile_prefix::String: prefix of the result-filename

logfile::String: name of logfile

exportflag::Bool: disable export for a simulation where saved results are not needed

ExportSettings([path::String, freq::Int])

Create ExportSettings only by path and freq. If no arguments are specified, the exportflag will be set to false and export disabled.

Arguments

path::String: path where results will be saved

freq::Int: export frequency

Peridynamics.BodySetup — Type

BodySetup

Setup of multiple bodies for PDContactAnalysis.

Fields:

pc::PointCloud: point cloud

mat::AbstractPDMaterial: material model

precracks::Vector{PreCrack}: predefined cracks

bcs::Vector{<:AbstractBC}: boundary conditions

ics::Vector{<:AbstractIC}: initial conditions

calc_timestep::Bool: use body for time step calculation

Peridynamics.Contact — Type

Contact

Contact definition.

Fields

body_id_set::Tuple{Int,Int}: bodies for which contact is defined

search_radius::Float64: search radius for contact definition to activate

spring_constant::Float64: spring constant used for contact force calculation, default value: spring_constant = 1.0e12

Peridynamics.PDSingleBodyAnalysis — Type

struct PDSingleBodyAnalysis{M, T} <:
    AbstractPDAnalysis where {M <: AbstractPDMaterial, T <: AbstractTimeDiscretization}

Peridynamic single body analysis.

Fields

name::String: simulation name

pc::PointCloud: point cloud

mat::M: material model for the body

precracks::Vector{PreCrack}: predefined cracks

bcs::Vector{<:AbstractBC}: boundary conditions

ics::Vector{<:AbstractIC}: initial conditions

td::T: time discretization

es::ExportSettings: export settings

Peridynamics.PDContactAnalysis — Type

PDContactAnalysis <: AbstractPDAnalysis

Peridynamic contact analysis.

Fields

name::String: simulation name

body_setup::Vector{BodySetup}: bodies used in this simulation

n_bodies::Int: number of bodies

contact::Vector{Contact}: contact definitions

td::VelocityVerlet: time discretization

es::ExportSettings: export settings

Functions

Peridynamics.pcmerge — Function

pcmerge(v::Vector{PointCloud})

Merge multiple point clouds into one PointCloud.

Arguments

v::Vector{PointCloud}: vector of multiple PointClouds

Returns

PointCloud: merged point cloud

Peridynamics.AbaqusMeshConverter.read_inp — Function

read_inp(file::String)

Read Abaqus .inp-file and convert meshes to PointCloud objects with the help of the AbaqusReader.jl package. Every element is converted to a point. The center of the element becomes the position of the point and the element volume becomes the point volume. Element sets defined in Abaqus are converted to corresponding point sets.

Currently supported mesh elements: [:Tet4, :Hex8]

Arguments

file::String: path to Abaqus .inp-file

Returns

PointCloud: generated point cloud with element volume as point volume

Examples

The specimen of the TensileTest.jl example script:

julia> pointcloud = read_inp("examples/models/TensileTestMesh.inp")
[ Info: 21420 nodes found
[ Info: Parsing elements. Type: C3D8. Topology: Hex8
[ Info: Creating elset bottom
[ Info: Creating elset top
16900-points PointCloud

julia> pointcloud.point_sets
Dict{String, Vector{Int64}} with 2 entries:
  "bottom" => [11701, 11702, 11703, 11704, 11705, 11706, 11707, 11708, 117…
  "top"    => [7501, 7502, 7503, 7504, 7505, 7506, 7507, 7508, 7509, 7510 …

Peridynamics.VtkReader.read_vtk — Function

read_vtk(file::String)

Read .vtu-file containing simulation results of a time step.

Arguments

file::String: path to VTK .vtu-file

Returns

Dict{String, VecOrMat{Float64}}: simulation results as a dictionary

Examples

julia> read_vtk("ExampleSimulation_t3000.vtu")
Dict{String, VecOrMat{Float64}} with 6 entries:
  "Position"     => [-0.497302 -0.497303 … 0.497303 0.497302; -0.0225023 -0…
  "Damage"       => [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0…
  "Time"         => [0.000237879]
  "Displacement" => [0.00019766 0.000196793 … -0.000196793 -0.00019766; -2.…
  "Velocity"     => [-2.55436 -1.5897 … 1.5897 2.55436; 0.827107 0.234996 ……
  "ForceDensity" => [-7.00131e9 -8.45411e9 … 8.45411e9 7.00131e9; -5.0164e8…

Peridynamics.submit — Function

submit(sim::T) where {T<:AbstractPDAnalysis}

Submit a simulation job to determine the results of the specified analysis.

Arguments

sim::T where {T<:AbstractPDAnalysis}: simulation job