Public API

Here, all exported types and methods are listed, which the user needs when using the package.

Peridynamics.BBMaterial

Peridynamics.Body

Peridynamics.CKIMaterial

Peridynamics.Job

Peridynamics.MultibodySetup

Peridynamics.NOSBMaterial

Peridynamics.OSBMaterial

Peridynamics.VelocityVerlet

Peridynamics.VtkReader.read_vtk

Peridynamics.contact!

Peridynamics.failure_permit!

Peridynamics.forcedensity_bc!

Peridynamics.material!

Peridynamics.point_set!

Peridynamics.precrack!

Peridynamics.submit

Peridynamics.uniform_box

Peridynamics.velocity_bc!

Peridynamics.velocity_ic!

Types

Peridynamics.BBMaterial — Type

BBMaterial <: AbstractMaterial

material type for bond-based peridynamic simulations

Allowed material parameters

:horizon::Float64: radius of point interactions

:rho::Float64: density

:E::Float64: Young's modulus

:Gc::Float64: critical energy release rate

:epsilon_c::Float64: critical strain

Allowed export fields

position::Matrix{Float64}: position of each point

displacement::Matrix{Float64}: displacement of each point

velocity::Matrix{Float64}: velocity of each point

velocity_half::Matrix{Float64}: velocity parameter for Verlet time solver

acceleration::Matrix{Float64}: acceleration of each point

b_int::Matrix{Float64}: internal force density of each point

b_ext::Matrix{Float64}: external force density of each point

damage::Vector{Float64}: damage of each point

n_active_bonds::Vector{Int}: number of intact bonds for each point

Peridynamics.OSBMaterial — Type

OSBMaterial <: AbstractMaterial

material type for ordinary state-based peridynamic simulations

Allowed material parameters

:horizon::Float64: radius of point interactions

:rho::Float64: density

:E::Float64: Young's modulus

:nu::Float64: Poisson's ratio

:Gc::Float64: critical energy release rate

:epsilon_c::Float64: critical strain

Allowed export fields

position::Matrix{Float64}: position of each point

displacement::Matrix{Float64}: displacement of each point

velocity::Matrix{Float64}: velocity of each point

velocity_half::Matrix{Float64}: velocity parameter for Verlet time solver

acceleration::Matrix{Float64}: acceleration of each point

b_int::Matrix{Float64}: internal force density of each point

b_ext::Matrix{Float64}: external force density of each point

damage::Vector{Float64}: damage of each point

n_active_bonds::Vector{Int}: number of intact bonds for each point

Peridynamics.NOSBMaterial — Type

NOSBMaterial <: AbstractMaterial

material type for non-ordinary state-based peridynamic simulations

Allowed material parameters

:horizon::Float64: radius of point interactions

:rho::Float64: density

:E::Float64: Young's modulus

:nu::Float64: Poisson's ratio

:Gc::Float64: critical energy release rate

:epsilon_c::Float64: critical strain

Allowed export fields

position::Matrix{Float64}: position of each point

displacement::Matrix{Float64}: displacement of each point

velocity::Matrix{Float64}: velocity of each point

velocity_half::Matrix{Float64}: velocity parameter for Verlet time solver

acceleration::Matrix{Float64}: acceleration of each point

b_int::Matrix{Float64}: internal force density of each point

b_ext::Matrix{Float64}: external force density of each point

damage::Vector{Float64}: damage of each point

n_active_bonds::Vector{Int}: number of intact bonds for each point

Peridynamics.CKIMaterial — Type

CKIMaterial <: AbstractMaterial

material type for continuum-kinematics-inspired peridynamic simulations

Allowed material parameters

:horizon::Float64: radius of point interactions

:rho::Float64: density

:E::Float64: Young's modulus

:nu::Float64: Poisson's ratio

:Gc::Float64: critical energy release rate

:epsilon_c::Float64: critical strain

Allowed export fields

TODO struct

Peridynamics.Body — Type

Body{M<:AbstractMaterial,P<:AbstractPointParameters}

Body for use in peridynamic calculation

Fields

mat<:AbstractMaterial: specified material model

n_points::Int: number of material points that represent the body

position::Matrix{Float64}: 3×n_points matrix with position for each point

volume::Vector{Float64}: vector with volume for each point

fail_permit::Vector{Bool}: vector that describes if failure is allowed for each point

point_sets::Dict{Symbol,Vector{Int}}: dictionary containing the defined point sets

point_params::Vector{P}: vector with material parameter sets

params_map::Vector{Int}: vector that assigns a material parameter set to each point

single_dim_bcs::Vector{SingleDimBC}: vector with defined boundary conditions

single_dim_ics::Vector{SingleDimIC}: vector with defined initial conditions

point_sets_precracks::Vector{PointSetsPreCrack}: vector with defined cracks

Constructors:

Body(mat<:AbstractMaterial, position::AbstractMatrix, volume::AbstractVector)

Creates a body for use in peridynamic calculation

Arguments

mat<:AbstractMaterial: specifies which material model is used

position::AbstractMatrix: 3×n matrix with position of each point

volume::AbstractVector: vector with volume of each point

Throws

error if n_points = 0

DimensionMismatch: error if dimension of position != 3

Example

julia> l, Δx, = 1.0, 1/50;
julia> pos, vol = uniform_box(l, l, 0.1l, Δx);
julia> b = Body(BBMaterial(), pos, vol);
julia> b
Body{BBMaterial, Peridynamics.BBPointParameters}(BBMaterial(), 12500,
[-0.49 -0.47 … 0.47 0.49; -0.49 -0.49 … 0.49 0.49; -0.04 -0.04 … 0.04 0.04],
[8.000000000000001e-6, 8.000000000000001e-6,  …  8.000000000000001e-6],
Bool[1, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
Dict{Symbol, Vector{Int64}}(), Peridynamics.BBPointParameters[],
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
Peridynamics.SingleDimBC[], Peridynamics.SingleDimIC[], Peridynamics.PointSetsPreCrack[])

Peridynamics.MultibodySetup — Type

MultibodySetup{M<:AbstractMaterial,P<:AbstractPointParameters}

setup for a peridynamic simulation with multiple bodies

Fields

bodies::Dict{Symbol,Body{M,P}}:

contacts::Vector{Contact}:

Constructors:

MultibodySetup(bodies::Dict{Symbol,Body{M,P}}) where {M,P}

Arguments

bodies::Dict{Symbol,Body{M,P}}:

Throws

error if less than 2 bodies are defined

error if defined bodies have different material types

Example

TODO

Peridynamics.VelocityVerlet — Type

VelocityVerlet <: AbstractTimeSolver

Procedure for calculating discrete time steps

Fields

end_time::Float64: time covered by the simulation

n_steps::Int: number of time steps calculated

Δt::Float64: size of discrete time steps

safety_factor::Float64: safety factor for step size to ensure stability

Constructors:

VelocityVerlet <: AbstractTimeSolver

Keywords

time::Real=-1: time covered by the simulation

steps::Int=-1: number of time steps calculated

stepsize::Real=-1: size of discrete time steps

safety_factor::Real=0.7: safety factor for step size to ensure stability

Throws

error if time and number of steps are specified

error if no time or number of steps are specified

error if safety factor is not between 0 and 1

Example

julia> vv = VelocityVerlet(steps=2000)

Peridynamics.Job — Type

Job{S<:AbstractSpatialSetup,T<:AbstractTimeSolver}

Job that contains all the information required for a peridynamic simulation

Fields

spatial_setup<:AbstractSpatialSetup: Body or Multibody setup for the simulation

time_solver<:AbstractTimeSolver: method for calculating discrete time steps

options::ExportOptions: settings for simulation data export

Constructors:

Job(spatial_setup<:AbstractSpatialSetup, time_solver<:AbstractTimeSolver,
    options::ExportOptions)
Job(spatial_setup<:AbstractSpatialSetup, time_solver<:AbstractTimeSolver; kwargs...)

Arguments

spatial_setup<:AbstractSpatialSetup: Body or Multibody setup for the simulation

time_solver<:AbstractTimeSolver: method for calculating discrete time steps

options::ExportOptions: settings for simulation data export

Keywords

:path::String: storage path for results

:freq::Int: frequency of time steps that are exported

:fields: exported fields Possible export fields depend on selected material model. See material type documentation. default export fields: :displacement, :damage

Throws

error if keyword is not allowed

Example

julia> b = Body(BBMaterial(), pos, vol)
julia> vv = VelocityVerlet(steps=2000)
julia> job = Job(b, vv;
           :path=joinpath(@__DIR__, "results", "mode_I"),
           :fields=(:displacement, :velocity, :acceleration, :damage))
Job{Body{BBMaterial, Peridynamics.BBPointParameters}, VelocityVerlet}(Body{BBMaterial,
Peridynamics.BBPointParameters}(BBMaterial(), 12500, [-0.49 -0.47 …

Functions

Peridynamics.point_set! — Function

point_set!(b::Body, name::Symbol, points::V) where {V<:AbstractVector}
point_set!(f::F, b::Body, name::Symbol) where {F<:Function}

Creates the point set name in Body b containing all points either defined in V or described by F

Arguments

b::Body: peridynamic body

name::Symbol: name of the point set

points<:AbstractVector: vector of point indices

f<:Function: function that describes points contained in point set

Throws

error if a point set called name is already defined

BoundsError: if points in V do not exist

Example

julia> point_set!(p -> p[1] ≤ -l/2+a && 0 ≤ p[2] ≤ 2δ, b, :set_a)
julia> point_set!(p -> p[1] ≤ -l/2+a && -2δ ≤ p[2] < 0, b, :set_b)
julia> b.point_sets
Dict{Symbol, Vector{Int64}} with 4 entries:
  :set_a      => [1251, 1252, 1253, 1254, 1255, 1256, 1257, 1258, 1259, …
  :set_b      => [951, 952, 953, 954, 955, 956, 957, 958, 959, 960  …  1…

Peridynamics.failure_permit! — Function

failure_permit!(b::Body, fail_permit::Bool)
failure_permit!(b::Body, name::Symbol, fail_permit::Bool)

determines whether failure is permitted fail_permit = true or prohibited fail_permit = false in the body b or the point set name of body b

Arguments

b::Body: peridynamic body

name::Symbol: name of the point set on body b

fail_permit::Bool: decides if failure is allowed on considered body or point set

Throws

error if no point set called name exists

Example

julia> failure_permit!(b, :set_bottom, false)
julia> b.fail_permit
12500-element Vector{Bool}:
 0
 0
 0
 ⋮
 1
 1

Peridynamics.material! — Function

material!(b::Body{M,P}, name::Symbol; kwargs...)
material!(b::Body{M,P}; kwargs...)
    where {M<:AbstractMaterial,P<:AbstractPointParameters}

specifies material parameters used for body b or point set name

Arguments

b::Body{M<:AbstractMaterial,P<:AbstractPointParameters}: peridynamic body

name::Symbol: point set on body b

Keywords

Allowed keywords depend on selected material model. See material type documentation. Default material parameter keywords:

:horizon::Float64: radius of point interactions

:rho::Float64: density

:E::Float64: Young's modulus

:nu::Float64: Poisson's ratio

:Gc::Float64: critical energy release rate

:epsilon_c::Float64: critical strain

Throws

error if parameter is not eligible for specification in selected material model

Example

julia> material!(b; horizon=δ, E=2.1e5, rho=8e-6, Gc=2.7)
julia> b.point_params
1-element Vector{Peridynamics.BBPointParameters}:
 Peridynamics.BBPointParameters(0.0603, 8.0e-6, 210000.0, 0.25, 84000.0, 140000.0, 84000.0,
84000.0, 2.7, 0.013329779199368195, 6.0671037207022026e10)

Peridynamics.velocity_bc! — Function

velocity_bc!(f::F, b::Body, name::Symbol, d::DimensionSpec) where {F<:Function}

specifies velocity boundary conditions for point set name on body b

Arguments

f<:Function: velocity condition function

b::Body: peridynamic body

name::Symbol: point set on body b

d::DimensionSpec: direction of velocity

Throws

error if no point set called name exists

error if dimension is not correctly specified

error if function is not suitable as condition function

Example

julia> velocity_bc!(t -> -9.81 * t, b, :set_bottom, :y)
julia> b.single_dim_bcs
1-element Vector{Peridynamics.SingleDimBC}:
 Peridynamics.SingleDimBC{var"#15#16"}(var"#15#16"(), :velocity_half, :set_bottom, 0x02)
julia> velocity_bc!(t -> 40, b, :set_a, 1)
julia> b.single_dim_bcs
2-element Vector{Peridynamics.SingleDimBC}:
 Peridynamics.SingleDimBC{var"#15#16"}(var"#15#16"(), :velocity_half, :set_bottom, 0x02)
 Peridynamics.SingleDimBC{var"#17#18"}(var"#17#18"(), :velocity_half, :set_a, 0x01)

Peridynamics.velocity_ic! — Function

velocity_ic!(b::Body, name::Symbol, d::DimensionSpec, value::Real)

specifies initital conditions for the velocity of points in point set name on body b

Arguments

b::Body: peridynamic body

name::Symbol: point set on body b

d::DimensionSpec: direction of velocity

value::Real: initial velocity value

Throws

error if no point set called name exists

error if dimension is not correctly specified

Example

julia> velocity_ic!(b, :set_b, :y, 20)
julia> b.single_dim_ics
1-element Vector{Peridynamics.SingleDimIC}:
 Peridynamics.SingleDimIC(20.0, :velocity, :set_b, 0x02)

Peridynamics.forcedensity_bc! — Function

forcedensity_bc!(f::F, b::Body, name::Symbol, d::DimensionSpec) where {F<:Function}

specifies boundary conditions for force density on points of point set name on body b

Arguments

f<:Function: condition function

b::Body: peridynamic body

name::Symbol: point set on body b

d::DimensionSpec: direction of force density

Throws

error if no point set called name exists

error if dimension is not correctly specified

error if function is not suitable as condition function

Example

julia> forcedensity_bc!(t -> 40, b, :set_a, 1)
julia> b.single_dim_bcs
1-element Vector{Peridynamics.SingleDimBC}:
 Peridynamics.SingleDimBC{var"#25#26"}(var"#25#26"(), :b_ext, :set_a, 0x01)

Peridynamics.precrack! — Function

precrack!(b::Body, set_a::Symbol, set_b::Symbol)

creates a crack between two point sets by prohibiting interaction between points of different point sets

Arguments

b::Body: peridynamic body

set_a::Symbol: first point set

set_b::Symbol: second point set

Throws

error if point set set_a or set_b does not exist

error if point sets contain common points

Example

julia> precrack!(b, :set_a, :set_b)
julia> b.point_sets_precracks
1-element Vector{Peridynamics.PointSetsPreCrack}:
 Peridynamics.PointSetsPreCrack(:set_a, :set_b)

Peridynamics.contact! — Function

contact!(ms::MultibodySetup, body_a::Symbol, body_b::Symbol; kwargs...)

defines contact between multiple bodies

Arguments

ms::MultibodySetup: the multibody setup defined to simulate the contact

body_a::Symbol: first body in contact

body_b::Symbol: second body in contact

Keywords

:radius::Float64:

:sc::Float64:

Throws

error if a called body is not defined in the multibody setup

error if keyword is not allowed

TODO kwargs

Peridynamics.VtkReader.read_vtk — Function

read_vtk(file::AbstractString)

Read vtu or pvtu file containing simulation results of a time step.

Arguments

file::String: path to VTK file in vtu or pvtu format

Returns

Dict{String, VecOrMat{Float64}}: simulation results as a dictionary

Examples

julia> read_vtk("ExampleSimulation_t3000.vtu")
Dict{String, VecOrMat{Float64}} with 6 entries:
  "position"     => [-0.497302 -0.497303 … 0.497303 0.497302; -0.0225023 -0…
  "damage"       => [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0…
  "time"         => [0.000237879]
  "displacement" => [0.00019766 0.000196793 … -0.000196793 -0.00019766; -2.…
  "velocity"     => [-2.55436 -1.5897 … 1.5897 2.55436; 0.827107 0.234996 ……
  "b_int"        => [-7.00131e9 -8.45411e9 … 8.45411e9 7.00131e9; -5.0164e8…

Peridynamics.uniform_box — Function

uniform_box(lx::Real, ly::Real, lz::Real, Δx::Real;
            center_x::Real=0, center_y::Real=0, center_z::Real=0)

Creates a grid of points in a cuboid form with lengths lx, ly and lz and the distance Δx between the points.

Arguments

lx::Real: length in x-dimension

ly::Real: length in y-dimension

lz::Real: length in z-dimension

Δx::Real: distance between neighboring points

Keywords

center_x::Real=0: center of the cuboid in x-dimension

center_y::Real=0: center of the cuboid in y-dimension

center_z::Real=0: center of the cuboid in z-dimension

Returns

position::Matrix{<:Real}: 3×n matrix with position of each point

volume::Vector{<:Real}: vector with volume of each point

Example

julia> position, volume = uniform_box(10, 10, 10, 2);

julia> position
3×125 Matrix{Float64}:
 -4.0  -2.0   0.0   2.0   4.0  -4.0  -2.0   0.0  …  0.0  2.0  4.0  -4.0  -2.0  0.0  2.0  4.0
 -4.0  -4.0  -4.0  -4.0  -4.0  -2.0  -2.0  -2.0     2.0  2.0  2.0   4.0   4.0  4.0  4.0  4.0
 -4.0  -4.0  -4.0  -4.0  -4.0  -4.0  -4.0  -4.0     4.0  4.0  4.0   4.0   4.0  4.0  4.0  4.0

julia> volume
125-element Vector{Int64}:
 8
 8
 8
 8
 ⋮
 8
 8
 8
 8

Peridynamics.submit — Function

submit(job::Job; kwargs...)

submits the job to start calculations

Arguments

job::Job: job that contains all defined parameters and conditions

Keywords

:quiet::Bool: if :quiet=true, no outputs are printed in the terminal

Throws

error if keyword is not allowed

Example

julia> job = Job(b, vv;
           path=joinpath(@__DIR__, "results", "mode_I"),
           fields=(:displacement, :damage))
julia> submit(job)
solve... 100%|████████████████████| Time: 0:00:11