Public API

Here, all exported types and methods are listed, which the user needs when using the package.

Peridynamics.BBMaterial

Peridynamics.Body

Peridynamics.CKIMaterial

Peridynamics.Job

Peridynamics.MultibodySetup

Peridynamics.NOSBMaterial

Peridynamics.OSBMaterial

Peridynamics.VelocityVerlet

Peridynamics.VtkReader.read_vtk

Peridynamics.contact!

Peridynamics.failure_permit!

Peridynamics.forcedensity_bc!

Peridynamics.material!

Peridynamics.point_set!

Peridynamics.precrack!

Peridynamics.submit

Peridynamics.uniform_box

Peridynamics.velocity_bc!

Peridynamics.velocity_ic!

Types

Peridynamics.BBMaterial — Type

BBMaterial <: AbstractMaterial

Material type for bond-based peridynamic simulations

Allowed material parameters

horizon::Float64: Radius of point interactions

rho::Float64: Density

E::Float64: Young's modulus

Gc::Float64: Critical energy release rate

epsilon_c::Float64: Critical strain

Allowed export fields

position::Matrix{Float64}: Position of each point

displacement::Matrix{Float64}: Displacement of each point

velocity::Matrix{Float64}: Velocity of each point

velocity_half::Matrix{Float64}: Velocity parameter for Verlet time solver

acceleration::Matrix{Float64}: Acceleration of each point

b_int::Matrix{Float64}: Internal force density of each point

b_ext::Matrix{Float64}: External force density of each point

damage::Vector{Float64}: Damage of each point

n_active_bonds::Vector{Int}: Number of intact bonds for each point

Peridynamics.OSBMaterial — Type

OSBMaterial <: AbstractMaterial

Material type for ordinary state-based peridynamic simulations

Allowed material parameters

horizon::Float64: Radius of point interactions

rho::Float64: Density

E::Float64: Young's modulus

nu::Float64: Poisson's ratio

Gc::Float64: Critical energy release rate

epsilon_c::Float64: Critical strain

Allowed export fields

position::Matrix{Float64}: Position of each point

displacement::Matrix{Float64}: Displacement of each point

velocity::Matrix{Float64}: Velocity of each point

velocity_half::Matrix{Float64}: Velocity parameter for Verlet time solver

acceleration::Matrix{Float64}: Acceleration of each point

b_int::Matrix{Float64}: Internal force density of each point

b_ext::Matrix{Float64}: External force density of each point

damage::Vector{Float64}: Damage of each point

n_active_bonds::Vector{Int}: Number of intact bonds for each point

Peridynamics.NOSBMaterial — Type

NOSBMaterial <: AbstractMaterial

Material type for non-ordinary state-based peridynamic simulations

Allowed material parameters

horizon::Float64: Radius of point interactions

rho::Float64: Density

E::Float64: Young's modulus

nu::Float64: Poisson's ratio

Gc::Float64: Critical energy release rate

epsilon_c::Float64: Critical strain

Allowed export fields

position::Matrix{Float64}: Position of each point

displacement::Matrix{Float64}: Displacement of each point

velocity::Matrix{Float64}: Velocity of each point

velocity_half::Matrix{Float64}: Velocity parameter for Verlet time solver

acceleration::Matrix{Float64}: Acceleration of each point

b_int::Matrix{Float64}: Internal force density of each point

b_ext::Matrix{Float64}: External force density of each point

damage::Vector{Float64}: Damage of each point

n_active_bonds::Vector{Int}: Number of intact bonds for each point

Peridynamics.CKIMaterial — Type

CKIMaterial <: AbstractMaterial

Material type for continuum-kinematics-inspired peridynamic simulations

Allowed material parameters

horizon::Float64: Radius of point interactions

rho::Float64: Density

E::Float64: Young's modulus

nu::Float64: Poisson's ratio

Gc::Float64: Critical energy release rate

epsilon_c::Float64: Critical strain

Allowed export fields

TODO struct

Peridynamics.Body — Type

Body(material, position, volume)

Creates a body for use in peridynamic calculation

Arguments

material::AbstractMaterial: Specifies which material model is used

position::AbstractMatrix: 3×n matrix with position of each point

volume::AbstractVector: Vector with volume of each point

Throws

Error if n_points = 0

DimensionMismatch: Error if dimension of position != 3

Example

julia> l, Δx, = 1.0, 1/50;

julia> pos, vol = uniform_box(l, l, 0.1l, Δx);

julia> b = Body(BBMaterial(), pos, vol);

julia> b
Body{BBMaterial, Peridynamics.BBPointParameters}(BBMaterial(), 12500,
[-0.49 -0.47 … 0.47 0.49; -0.49 -0.49 … 0.49 0.49; -0.04 -0.04 … 0.04 0.04],
[8.000000000000001e-6, 8.000000000000001e-6,  …  8.000000000000001e-6],
Bool[1, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
Dict{Symbol, Vector{Int64}}(), Peridynamics.BBPointParameters[],
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
Peridynamics.SingleDimBC[], Peridynamics.SingleDimIC[], Peridynamics.PointSetsPreCrack[])

Internal use only

Please note that the fields are intended for internal use only. They are not part of the public API of Peridynamics.jl, and thus can be altered (or removed) at any time without it being considered a breaking change.

Body{Material,PointParameters}

Type Parameters

Material <: AbstractMaterial: Type of the specified material model

PointParameters <: AbstractPointParameters: Type of the point parameters

Fields

mat::Material: Specified material model

n_points::Int: Number of material points that represent the body

position::Matrix{Float64}: 3×n_points matrix with position for each point

volume::Vector{Float64}: Vector with volume for each point

fail_permit::Vector{Bool}: Vector that describes if failure is allowed for each point

point_sets::Dict{Symbol,Vector{Int}}: Dictionary containing the defined point sets

point_params::Vector{PointParameters}: Vector with material parameter sets

params_map::Vector{Int}: Vector that assigns a material parameter set to each point

single_dim_bcs::Vector{SingleDimBC}: Vector with defined boundary conditions

single_dim_ics::Vector{SingleDimIC}: Vector with defined initial conditions

point_sets_precracks::Vector{PointSetsPreCrack}: Vector with defined cracks

Peridynamics.MultibodySetup — Type

MultibodySetup(bodies)

Setup for a peridynamic simulation with multiple bodies

Arguments

bodies::Dict{Symbol,Body{M,P}}:

Throws

Error if less than 2 bodies are defined

Error if defined bodies have different material types

Example

Internal use only

Please note that the fields are intended for internal use only. They are not part of the public API of Peridynamics.jl, and thus can be altered (or removed) at any time without it being considered a breaking change.

MultibodySetup{Material,PointParameters}

Type Parameters

Material <: AbstractMaterial: Type of the specified material model

PointParameters <: AbstractPointParameters: Type of the point parameters

Fields

bodies::Dict{Symbol,Body{M,P}}:

contacts::Vector{Contact}:

TODO

Peridynamics.VelocityVerlet — Type

VelocityVerlet(; time=-1, steps=-1, stepsize=-1, safety_factor=0.7)

Procedure for calculating discrete time steps

Keywords

time::Real=-1: Time covered by the simulation

steps::Int=-1: Number of calculated time steps

stepsize::Real=-1: Size of discrete time steps

safety_factor::Real=0.7: Safety factor for step size to ensure stability

Throws

Error if time and number of steps are specified

Error if no time or number of steps are specified

Error if safety factor is not between 0 and 1

Example

julia> vv = VelocityVerlet(steps=2000)

Internal use only

Please note that the fields are intended for internal use only. They are not part of the public API of Peridynamics.jl, and thus can be altered (or removed) at any time without it being considered a breaking change.

VelocityVerlet <: AbstractTimeSolver

Type Parameter

-AbstractTimeSolver: Type of the time solver

Fields

end_time::Float64: Time covered by the simulation

n_steps::Int: Number of calculated time steps

Δt::Float64: Size of discrete time steps

safety_factor::Float64: Safety factor for step size to ensure stability

Peridynamics.Job — Type

Job(spatial_setup, time_solver; kwargs...)

Job that contains all the information required for a peridynamic simulation

Arguments

spatial_setup::AbstractSpatialSetup: Body or Multibody setup for the simulation

time_solver::AbstractTimeSolver: Method for calculating discrete time steps

Keywords

path::String: Storage path for results

freq::Int: Frequency of time steps that are exported

fields::NTuple{N,Symbol}: Exported fields Possible export fields depend on the selected material model. See material type documentation. Default export fields: (:displacement, :damage)

Throws

Error if keyword is not allowed

Example

julia> b = Body(BBMaterial(), pos, vol)

julia> vv = VelocityVerlet(steps=2000)

julia> job = Job(b, vv;
           path=joinpath(@__DIR__, "results", "mode_I"),
           fields=(:displacement, :velocity, :acceleration, :damage))
Job{Body{BBMaterial, Peridynamics.BBPointParameters}, VelocityVerlet}(Body{BBMaterial,
Peridynamics.BBPointParameters}(BBMaterial(), 12500, [-0.49 -0.47 …

Note that for the keyword fields a NTuple is expected! If you want to export only one field, insert , after the field:

julia> job = Job(b, vv;
           path=joinpath(@__DIR__, "results", "mode_I"),
           fields=(:displacement,))

Internal use only

Please note that the fields are intended for internal use only. They are not part of the public API of Peridynamics.jl, and thus can be altered (or removed) at any time without it being considered a breaking change.

Job{SpatialSetup,TimeSolver,Options}

Type Parameters

SpatialSetup <: AbstractSpatialSetup: Type of the spatial setup

TimeSolver <: AbstractTimeSolver: Type of the time solver

Options <: AbstractOptions: Type of the export options

Fields

spatial_setup::AbstractSpatialSetup: Body or Multibody setup for the simulation

time_solver::AbstractTimeSolver: Method for calculating discrete time steps

options::AbstractOptions: Options for simulation data export

Functions

Peridynamics.point_set! — Function

point_set!(body, set, points)
point_set!(fun, body, set)

Creates the point set set in body containing all points either defined in vector points or described by function fun.

Arguments

body::AbstractBody: Peridynamic body

set::Symbol: Point set on body

points::AbstractVector: Vector of point indices

fun::Function: Function that describes points contained in point set

Throws

Error if a point set called set is already defined

BoundsError: If points in V do not exist

Example

julia> point_set!(p -> p[1] ≤ -l/2+a && 0 ≤ p[2] ≤ 2δ, b, :set_a)

julia> point_set!(p -> p[1] ≤ -l/2+a && -2δ ≤ p[2] < 0, b, :set_b)

julia> b.point_sets
Dict{Symbol, Vector{Int64}} with 4 entries:
  :set_a      => [1251, 1252, 1253, 1254, 1255, 1256, 1257, 1258, 1259, …
  :set_b      => [951, 952, 953, 954, 955, 956, 957, 958, 959, 960  …  1…

Peridynamics.failure_permit! — Function

failure_permit!(body, fail_permit)
failure_permit!(body, set, fail_permit)

Determines whether failure is permitted fail_permit = true or prohibited fail_permit = false in the body or the point set set of the body.

Arguments

body::AbstractBody: Peridynamic body

set::Symbol: Point set on body

fail_permit::Bool: Decides if failure is allowed on considered body or point set.

Throws

Error if no point set called set exists

Example

julia> failure_permit!(b, :set_bottom, false)

julia> b.fail_permit
12500-element Vector{Bool}:
 0
 0
 0
 ⋮
 1
 1

Peridynamics.material! — Function

material!(body, set; kwargs...)
material!(body; kwargs...)

Specifies material parameters used for body or point set set

Arguments

body::AbstractBody: Peridynamic body

set::Symbol: Point set on body

Keywords

Allowed keywords depend on the selected material model. See material type documentation. Default material parameter keywords:

horizon::Float64: Radius of point interactions

rho::Float64: Density

E::Float64: Young's modulus

nu::Float64: Poisson's ratio

Gc::Float64: Critical energy release rate

epsilon_c::Float64: Critical strain

Throws

Error if parameter is not eligible for specification in selected material model

Example

julia> material!(b; horizon=δ, E=2.1e5, rho=8e-6, Gc=2.7)

julia> b.point_params
1-element Vector{Peridynamics.BBPointParameters}:
 Peridynamics.BBPointParameters(0.0603, 8.0e-6, 210000.0, 0.25, 84000.0, 140000.0, 84000.0,
84000.0, 2.7, 0.013329779199368195, 6.0671037207022026e10)

Peridynamics.velocity_bc! — Function

velocity_bc!(fun, body, set, dim)

Specifies velocity boundary conditions for point set set on body

Arguments

fun::Function: Velocity condition function

body::AbstractBody: Peridynamic body

set::Symbol: Point set on body

dim::Union{Integer,Symbol}: Direction of velocity

Throws

Error if no point set called set exists

Error if dimension is not correctly specified

Error if function is not suitable as condition function

Example

julia> velocity_bc!(t -> -9.81 * t, b, :set_bottom, :y)

julia> b.single_dim_bcs

1-element Vector{Peridynamics.SingleDimBC}:
 Peridynamics.SingleDimBC{var"#15#16"}(var"#15#16"(), :velocity_half, :set_bottom, 0x02)

julia> velocity_bc!(t -> 40, b, :set_a, 1)

julia> b.single_dim_bcs
2-element Vector{Peridynamics.SingleDimBC}:
 Peridynamics.SingleDimBC{var"#15#16"}(var"#15#16"(), :velocity_half, :set_bottom, 0x02)
 Peridynamics.SingleDimBC{var"#17#18"}(var"#17#18"(), :velocity_half, :set_a, 0x01)

Peridynamics.velocity_ic! — Function

velocity_ic!(body, set, dim, value)

Specifies initital conditions for the velocity of points in point set set on body

Arguments

body::AbstractBody: Peridynamic body

set::Symbol: Point set on body

dim::Union{Integer,Symbol}: Direction of velocity

value::Real: Initial velocity value

Throws

Error if no point set called set exists

Error if dimension is not correctly specified

Example

julia> velocity_ic!(b, :set_b, :y, 20)

julia> b.single_dim_ics
1-element Vector{Peridynamics.SingleDimIC}:
 Peridynamics.SingleDimIC(20.0, :velocity, :set_b, 0x02)

Peridynamics.forcedensity_bc! — Function

forcedensity_bc!(fun, body, set, dim)

Specifies boundary conditions for force density on points of point set set on body

Arguments

fun::Function: Condition function

body::AbstractBody: Peridynamic body

set::Symbol: Point set on body

dim::Union{Integer,Symbol}: Direction of force density

Throws

Error if no point set called set exists

Error if dimension is not correctly specified

Error if function is not suitable as condition function

Example

julia> forcedensity_bc!(t -> 40, b, :set_a, 1)

julia> b.single_dim_bcs
1-element Vector{Peridynamics.SingleDimBC}:
 Peridynamics.SingleDimBC{var"#25#26"}(var"#25#26"(), :b_ext, :set_a, 0x01)

Peridynamics.precrack! — Function

precrack!(body, set_a, set_b; update_dmg=true)

Creates a crack between two point sets by prohibiting interaction between points of different point sets.

Arguments

body::AbstractBody: Peridynamic body

set_a::Symbol: First point set on body

set_b::Symbol: Second point set on body

Keywords

update_dmg::Bool: If update_dmg=true, the material points involved in the predefined are initially damaged. If update_dmg=false, the bonds involved are deleted and the material points involved with the predefined crack are not damaged. (default=true)

Throws

Error if point set set_a or set_b does not exist

Error if point sets contain common points

Example

julia> precrack!(b, :set_a, :set_b)

julia> b.point_sets_precracks
1-element Vector{Peridynamics.PointSetsPreCrack}:
 Peridynamics.PointSetsPreCrack(:set_a, :set_b)

Peridynamics.contact! — Function

contact!(ms, body_a, body_b; kwargs...)

Defines contact between multiple bodies

Arguments

ms::AbstractMultibodySetup: The multibody setup defined to simulate the contact

body_a::Symbol: First body in contact

body_b::Symbol: Second body in contact

Keywords

radius::Float64:

sc::Float64:

Throws

Error if a called body is not defined in the multibody setup

Error if keyword is not allowed

TODO kwargs

Peridynamics.VtkReader.read_vtk — Function

read_vtk(file::AbstractString)

Read vtu or pvtu file containing simulation results of a time step.

Arguments

file::String: Path to VTK file in vtu or pvtu format

Returns

Dict{String, VecOrMat{Float64}}: Simulation results as a dictionary

Examples

julia> read_vtk("ExampleSimulation_t3000.vtu")
Dict{String, VecOrMat{Float64}} with 6 entries:
  "position"     => [-0.497302 -0.497303 … 0.497303 0.497302; -0.0225023 -0…
  "damage"       => [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0…
  "time"         => [0.000237879]
  "displacement" => [0.00019766 0.000196793 … -0.000196793 -0.00019766; -2.…
  "velocity"     => [-2.55436 -1.5897 … 1.5897 2.55436; 0.827107 0.234996 ……
  "b_int"        => [-7.00131e9 -8.45411e9 … 8.45411e9 7.00131e9; -5.0164e8…

Peridynamics.uniform_box — Function

uniform_box(lx, ly, lz, Δx; center_x=0, center_y=0, center_z=0)

Creates a grid of points in a cuboid form with lengths lx, ly and lz and the distance Δx between the points.

Arguments

lx::Real: Length in x-dimension

ly::Real: Length in y-dimension

lz::Real: Length in z-dimension

Δx::Real: Distance between neighboring points

Keywords

center_x::Real=0: Center of the cuboid in x-dimension

center_y::Real=0: Center of the cuboid in y-dimension

center_z::Real=0: Center of the cuboid in z-dimension

Returns

position::Matrix: 3×n matrix with position of each point

volume::Vector: Vector with volume of each point

Example

julia> position, volume = uniform_box(10, 10, 10, 2);

julia> position
3×125 Matrix{Float64}:
 -4.0  -2.0   0.0   2.0   4.0  -4.0  -2.0   0.0  …  0.0  2.0  4.0  -4.0  -2.0  0.0  2.0  4.0
 -4.0  -4.0  -4.0  -4.0  -4.0  -2.0  -2.0  -2.0     2.0  2.0  2.0   4.0   4.0  4.0  4.0  4.0
 -4.0  -4.0  -4.0  -4.0  -4.0  -4.0  -4.0  -4.0     4.0  4.0  4.0   4.0   4.0  4.0  4.0  4.0

julia> volume
125-element Vector{Int64}:
 8
 8
 8
 8
 ⋮
 8
 8
 8
 8

Peridynamics.submit — Function

submit(job::Job; kwargs...)

Submits the job to start calculations

Arguments

job::Job: Job that contains all defined parameters and conditions

Keywords

quiet::Bool: If :quiet=true, no outputs are printed in the terminal.

Throws

Error if keyword is not allowed

Example

julia> job = Job(b, vv;
           path=joinpath(@__DIR__, "results", "mode_I"),
           fields=(:displacement, :damage));

julia> submit(job);
solve... 100%|████████████████████| Time: 0:00:11