Private API

Private API

Peridynamics.failure_permit! — Function

failure_permit!(body, set_name, fail_permit)

Set the failure permission for points of the set set_name of a body.

Arguments

body::AbstractBody: Body where the failure permission will be set.

set_name::Symbol: The name of a point set of this body.

fail_permit::Bool: If true, failure is allowed, and if false then no bonds of this point are allowed to break during the simulation.

Overwriting failure permission with `material!` and `failure_permit!`

The function material! calls failure_permit!, so if it is used afterwards, previously set failure permissions might be overwritten!

Throws

Error if the body does not contain a set with set_name.